+++ src/Pickles/Pickles.TestFrameworks.UnitTests/WhenResolvingTestResults.cs
@@ -31,8 +31,11 @@ using PicklesDoc.Pickles.Test;
 using PicklesDoc.Pickles.TestFrameworks.CucumberJson;
 using PicklesDoc.Pickles.TestFrameworks.MsTest;
 using PicklesDoc.Pickles.TestFrameworks.NUnit.NUnit2;
+using PicklesDoc.Pickles.TestFrameworks.NUnit.NUnit3;
 using PicklesDoc.Pickles.TestFrameworks.SpecRun;
+using PicklesDoc.Pickles.TestFrameworks.VsTest;
 using PicklesDoc.Pickles.TestFrameworks.XUnit.XUnit1;
+using PicklesDoc.Pickles.TestFrameworks.XUnit.XUnit2;
 
 namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
 {
@@ -93,6 +96,25 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
         }
 
         [Test]
+        public void ThenCanResolveAsSingletonWhenTestResultsAreNUnit3()
+        {
+            FileSystem.AddFile("results-example-nunit3.xml", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "NUnit.NUnit3.results-example-nunit3.xml"));
+
+            var configuration = Container.Resolve<IConfiguration>();
+            configuration.TestResultsFormat = TestResultsFormat.NUnit3;
+            configuration.AddTestResultFile(FileSystem.FileInfo.FromFileName("results-example-nunit3.xml"));
+
+            var item1 = Container.Resolve<ITestResults>();
+            var item2 = Container.Resolve<ITestResults>();
+
+            Check.That(item1).IsNotNull();
+            Check.That(item1).IsInstanceOf<NUnit3Results>();
+            Check.That(item2).IsNotNull();
+            Check.That(item2).IsInstanceOf<NUnit3Results>();
+            Check.That(item1).IsSameReferenceThan(item2);
+        }
+
+        [Test]
         public void ThenCanResolveAsSingletonWhenTestResultsArexUnit()
         {
             FileSystem.AddFile("results-example-xunit.xml", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "XUnit.XUnit1.results-example-xunit.xml"));
@@ -108,6 +130,25 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
             Check.That(item1).IsInstanceOf<XUnit1Results>();
             Check.That(item2).IsNotNull();
             Check.That(item2).IsInstanceOf<XUnit1Results>();
+            Check.That(item1).IsSameReferenceThan(item2);
+        }
+
+        [Test]
+        public void ThenCanResolveAsSingletonWhenTestResultsArexUnit2()
+        {
+            FileSystem.AddFile("results-example-xunit2.xml", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "XUnit.XUnit2.results-example-xunit2.xml"));
+
+            var configuration = Container.Resolve<IConfiguration>();
+            configuration.TestResultsFormat = TestResultsFormat.xUnit2;
+            configuration.AddTestResultFile(FileSystem.FileInfo.FromFileName("results-example-xunit2.xml"));
+
+            var item1 = Container.Resolve<ITestResults>();
+            var item2 = Container.Resolve<ITestResults>();
+
+            Check.That(item1).IsNotNull();
+            Check.That(item1).IsInstanceOf<XUnit2Results>();
+            Check.That(item2).IsNotNull();
+            Check.That(item2).IsInstanceOf<XUnit2Results>();
             Check.That(item1).IsSameReferenceThan(item2);
         }
 
@@ -146,6 +187,25 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
             Check.That(item1).IsInstanceOf<SpecRunResults>();
             Check.That(item2).IsNotNull();
             Check.That(item2).IsInstanceOf<SpecRunResults>();
+            Check.That(item1).IsSameReferenceThan(item2);
+        }
+
+        [Test]
+        public void ThenCanResolveAsSingletonWhenTestResultsAreVsTest()
+        {
+            FileSystem.AddFile("results-example-vstest.trx", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "VsTest.results-example-vstest.trx"));
+
+            var configuration = Container.Resolve<IConfiguration>();
+            configuration.TestResultsFormat = TestResultsFormat.VsTest;
+            configuration.AddTestResultFiles(new[] { FileSystem.FileInfo.FromFileName("results-example-vstest.trx") });
+
+            var item1 = Container.Resolve<ITestResults>();
+            var item2 = Container.Resolve<ITestResults>();
+
+            Check.That(item1).IsNotNull();
+            Check.That(item1).IsInstanceOf<VsTestResults>();
+            Check.That(item2).IsNotNull();
+            Check.That(item2).IsInstanceOf<VsTestResults>();
             Check.That(item1).IsSameReferenceThan(item2);
         }
 
@@ -189,6 +249,21 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
         }
 
         [Test]
+        public void ThenCanResolveWhenTestResultsAreNUnit3()
+        {
+            FileSystem.AddFile("results-example-nunit3.xml", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "NUnit.NUnit3.results-example-nunit3.xml"));
+
+            var configuration = Container.Resolve<IConfiguration>();
+            configuration.TestResultsFormat = TestResultsFormat.NUnit3;
+            configuration.AddTestResultFile(FileSystem.FileInfo.FromFileName("results-example-nunit3.xml"));
+
+            var item1 = Container.Resolve<ITestResults>();
+
+            Check.That(item1).IsNotNull();
+            Check.That(item1).IsInstanceOf<NUnit3Results>();
+        }
+
+        [Test]
         public void ThenCanResolveWhenTestResultsArexUnit()
         {
             FileSystem.AddFile("results-example-xunit.xml", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "XUnit.XUnit1.results-example-xunit.xml"));
@@ -201,6 +276,21 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
 
             Check.That(item).IsNotNull();
             Check.That(item).IsInstanceOf<XUnit1Results>();
+        }
+
+        [Test]
+        public void ThenCanResolveWhenTestResultsArexUnit2()
+        {
+            FileSystem.AddFile("results-example-xunit2.xml", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "XUnit.XUnit2.results-example-xunit2.xml"));
+
+            var configuration = Container.Resolve<IConfiguration>();
+            configuration.TestResultsFormat = TestResultsFormat.xUnit2;
+            configuration.AddTestResultFile(FileSystem.FileInfo.FromFileName("results-example-xunit2.xml"));
+
+            var item1 = Container.Resolve<ITestResults>();
+
+            Check.That(item1).IsNotNull();
+            Check.That(item1).IsInstanceOf<XUnit2Results>();
         }
 
         [Test]
@@ -231,6 +321,21 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
 
             Check.That(item).IsNotNull();
             Check.That(item).IsInstanceOf<SpecRunResults>();
+        }
+
+        [Test]
+        public void ThenCanResolveWhenTestResultsAreVsTest()
+        {
+            FileSystem.AddFile("results-example-vstest.trx", RetrieveContentOfFileFromResources(TestResultsResourcePrefix + "VsTest.results-example-vstest.trx"));
+
+            var configuration = Container.Resolve<IConfiguration>();
+            configuration.TestResultsFormat = TestResultsFormat.VsTest;
+            configuration.AddTestResultFiles(new[] { FileSystem.FileInfo.FromFileName("results-example-vstest.trx") });
+
+            var item1 = Container.Resolve<ITestResults>();
+
+            Check.That(item1).IsNotNull();
+            Check.That(item1).IsInstanceOf<VsTestResults>();
         }
     }
 }